+++ src/main/java/com/owori/domain/member/dto/response/MemberProfileResponse.java
@@ -1,4 +1,4 @@
-package com.owori.domain.home.dto.response;
+package com.owori.domain.member.dto.response;
 
 import com.owori.domain.member.entity.EmotionalBadge;
 import lombok.AllArgsConstructor;
