+++ src/test/java/com/owori/domain/member/service/MemberServiceTest.java
@@ -1,12 +1,18 @@
 package com.owori.domain.member.service;
 
+import com.owori.domain.family.dto.request.FamilyRequest;
 import com.owori.domain.family.entity.Family;
+import com.owori.domain.family.repository.FamilyRepository;
+import com.owori.domain.family.service.FamilyService;
 import com.owori.domain.member.dto.request.EmotionalBadgeRequest;
 import com.owori.domain.member.dto.request.MemberDetailsRequest;
 import com.owori.domain.member.dto.request.MemberProfileRequest;
-import com.owori.domain.member.entity.Color;
-import com.owori.domain.member.entity.EmotionalBadge;
-import com.owori.domain.member.entity.Member;
+import com.owori.domain.member.dto.response.FindHomeResponse;
+import com.owori.domain.member.dto.response.MemberProfileResponse;
+import com.owori.domain.member.entity.*;
+import com.owori.domain.saying.dto.response.FindSayingByFamilyResponse;
+import com.owori.domain.saying.entity.Saying;
+import com.owori.domain.saying.repository.SayingRepository;
 import com.owori.global.exception.EntityNotFoundException;
 import com.owori.support.database.DatabaseTest;
 import com.owori.support.database.LoginTest;
@@ -16,6 +22,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 
 import javax.persistence.EntityManager;
 import java.time.LocalDate;
+import java.util.List;
+import java.util.Optional;
 import java.util.UUID;
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -26,6 +34,9 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 @DisplayName("Member 서비스의")
 class MemberServiceTest extends LoginTest {
     @Autowired private MemberService memberService;
+    @Autowired private FamilyService familyService;
+    @Autowired private FamilyRepository familyRepository;
+    @Autowired private SayingRepository sayingRepository;
     @Autowired private EntityManager em;
 
     @Test
@@ -134,5 +145,39 @@ class MemberServiceTest extends LoginTest {
 
         //then
         assertThat(authService.getLoginUser().getEmotionalBadge()).isEqualTo(soHappy).isNotEqualTo(before);
+    }
+
+    @Test
+    @DisplayName("홈 데이터 조회가 수행되는가")
+    void findHomeData() {
+        // given
+        // 가족 생성
+        String familyName = "오월이 가족";
+        String code = familyService.saveFamily(new FamilyRequest(familyName)).getInviteCode();
+
+        // 가족 구성원 생성
+        Member member1 = Member.builder().oAuth2Info(new OAuth2Info("111111", AuthProvider.APPLE)).build();
+        Member member2 = Member.builder().oAuth2Info(new OAuth2Info("222222", AuthProvider.APPLE)).build();
+        Member saveMember1 = memberRepository.save(member1);
+        Member saveMember2 = memberRepository.save(member2);
+
+        // 가족에 멤버 추가
+        Optional<Family> family = familyRepository.findByInviteCode(code);
+        if(family.isPresent()) {
+            family.get().addMember(saveMember1);
+            family.get().addMember(authService.getLoginUser());
+        }
+
+        // 서로에게 한마디 생성
+        Saying saying1 = sayingRepository.save(new Saying("오늘 집 안 감", authService.getLoginUser(), List.of()));
+        Saying saying2 = sayingRepository.save(new Saying("오늘 집 감", saveMember1, List.of()));
+        Saying saying3 = sayingRepository.save(new Saying("배고파", saveMember2, List.of()));
+
+        // when
+        FindHomeResponse responses = memberService.findHomeData();
+
+        assertThat(familyName).isEqualTo(responses.getFamilyName());
+        assertThat(responses.getMemberProfiles().stream().map(MemberProfileResponse::getId).toList()).isEqualTo(List.of(authService.getLoginUser().getId(), saveMember1.getId()));
+        assertThat(responses.getSayings().stream().map(FindSayingByFamilyResponse::getId)).hasSameElementsAs(List.of(saying1.getId(), saying2.getId()));
     }
 }
